+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Include a workspace_data_object context variable for the `WorkspaceDetail` and `WorkspaceUpdate` views.
 * Refactor auditing classes.
+* Add ability to specify a custom `Workspace` form in the workspace adapter.
 
 ## 0.17 (2023-07-11)